% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Allocation and categorization of climbs on the route}
\usage{
categorize(gps)
}
\arguments{
\item{gps}{the gps data of the cycling route (longitude, latitude, time, elevation, ...).}
}
\value{
A list containing:
  \itemize{
   \item gps the input data frame complemented with the smoothed elevation and hill grade and categorization of climbs (additional columns \code{ele_smooth}, \code{grade_smooth}, \code{category}),
   \item climbs data frame with details about all climbs with columns:
     \itemize{
       \item ind_start location of climb start (index of the input date frame),
       \item ind_end location of climb end (index of the input date frame),
       \item len length of the climb in meters,
       \item avg_grade average grade of the climb between the start and end points,
       \item vam vertical ascent in meters of each climb,
       \item category climb category (factorized),
     }
   \item total_vam total vertical ascent meters (VAM),
 }
}
\description{
This function determines the starting and ending point, length, average grade and category of climbs on a cycling route described with GPS data.
The categorization data is added to the gps data, while the details of the climbs are returned in a separate data frame.
Also added is the number of total vertical ascent meters (VAM) of the route.
}
\examples{
load(system.file("extdata", "gps.Rda", package="cycleR"))
hill_cat <- categorize(gps)
}
\keyword{GPS}
\keyword{analysis,}
\keyword{categorization}
\keyword{climb}
\keyword{cycling}
\keyword{data}
\keyword{power,}

