% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_member.R
\name{clean_member}
\alias{clean_member}
\title{Cleans member data stored in string variable}
\usage{
clean_member(
  df,
  col = "member",
  keep = FALSE,
  true = c("Member"),
  false = c("guest", "Casual", "Unknown")
)
}
\arguments{
\item{df}{the dataframe that contains ride-level data.}

\item{col}{the string column containing member information.}

\item{keep}{should the original member column be preserved in the output?}
}
\value{
a dataframe of all ammended member information in \code{member}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{clean_member} takes ride-level data in a data.frame and turns the member data stored in a string
variable into a binary column.
}
