% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{combine.two.BSJ.tables}
\alias{combine.two.BSJ.tables}
\title{combine BSJs}
\usage{
combine.two.BSJ.tables(ce_bsjs, ciri_bsjs)
}
\arguments{
\item{ce_bsjs}{BSJ table 1}

\item{ciri_bsjs}{BSJ table 2}
}
\value{
\code{Tibble} object with combined filtered BSJ coordinate and number of junction spanning 
 reads across sample.
}
\description{
Combine 2 BSJ tables
}
\details{
Just a combination of BSJ tables to make sure we have a complete set of BSJs. 
 The variable names do not actually matter since the all tables have the same formatting.
}
\author{
Stefan Stefanov
}
\keyword{BSJ}
\keyword{combine}
