% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycling_power.R
\name{cycling_power}
\alias{cycling_power}
\title{Assessment of the total power produced by a cyclist on a bike ride given the GPS data and additional physical parameters
The power required from a person to move on a bike between measured points of a GPS route is calculated}
\usage{
cycling_power(gps, m = 0, Crr = 0, Cd = 0, ro = 0, A = 0,
  windspeed = 0)
}
\arguments{
\item{gps}{the GPS data of the cycling route (longitude, latitude, time, elevation, ...),}

\item{m}{the combined mass of the rider and the bicycle (in kilograms),}

\item{Crr}{the rolling resistance coefficient defined based on the type of bike (road, mtb, cross) used,}

\item{Cd}{the drag coefficient (its value depends on the bike, equipment and rider position; typical values can be found in: Wilson, David Gordon, and Jim Papadopoulos. Bicycling science, p. 188. Mit Press, 2004.),}

\item{ro}{air density (in kg/m^3). At 25 degrees Celsius and 300 m elevation air density is approximately 1.13 kg/m^3,}

\item{A}{the surface area of the rider facing the wind (typical values can be found in: Wilson, David Gordon, and Jim Papadopoulos. Bicycling science, p. 188. Mit Press, 2004.),}

\item{windspeed}{average wind speed in m/s. Headwind is given as a negative number and tail wind is given as a positive number. This argument has value of 0 if there is no information about wind speed (default).}
}
\value{
A data frame with variables:
 \itemize{
   \item Proll the power required to overcome the rolling resistance,
   \item Pdrag the power required to overcome the wind resistance,
   \item Pclimb the power required to overcome the gravity pull,
   \item Pacc the power required to accelerate from one speed to another.
   \item Ptotal total cycling power for the given gps route.
 }
}
\description{
This function calculates the total power required from a person to move on a bike between measured points of a GPS route, based on the formula \eqn{P = Proll + Pdrag + Pclimb + Pacc}, where \eqn{Proll} is the power required to overcome the rolling resistance, \eqn{Pdrag} is the power required to overcome the wind resistance, \eqn{Pclimb} is the power required to overcome the pull of gravity and \eqn{Pacc} is the power required to accelerate from one speed to another.
}
\examples{
load(system.file("extdata", "gps.Rda", package="cycleR"))
Pcycling <- cycling_power(gps, Crr = 0.004, Cd = 0.9, ro = 1.13, A = 0.5, windspeed = 0)
}
\keyword{cycling}
\keyword{power}

