% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{filter_bam}
\alias{filter_bam}
\title{BAM file filter}
\usage{
filter_bam(BSJ_gr, sample_table, samtools_prefix)
}
\arguments{
\item{BSJ_gr}{a GRange of BSJ cooredinates}

\item{sample_table}{sample table formatted according to the manual,
Must contain \dQuote{sample_name} \dQuote{treatment} \dQuote{file_bam} \dQuote{lib_size} 
\dQuote{read_len}; NB the values in column \dQuote{treatment} can only be \dQuote{control} and 
\dQuote{enriched}}

\item{samtools_prefix}{a string that corresponds to user's samtools run prefix}
}
\value{
\code{BAMFileList} object with info on the trimmed files
}
\description{
A wrapper function for samtools use to trim the files
}
\details{
This function removes the BAM file reads that do not overlap with the BSJ loci.
 This significantly speeds us the feature detection and lowers the virtual memory requirements
}
\author{
Stefan Stefanov
}
\keyword{Bam}
\keyword{Filter}
