% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{find_depleted_features}
\alias{find_depleted_features}
\title{CircRNA feature selection}
\usage{
find_depleted_features(circ_fc_adj, sample_table, circ_sg, test = "DEX")
}
\arguments{
\item{circ_fc_adj}{count matrix corresponding to the circRNA features}

\item{sample_table}{sample table formatted according to the manual,
Must contain \dQuote{sample_name} \dQuote{treatment} \dQuote{file_bam} \dQuote{lib_size} 
\dQuote{read_len}; NB the values in column \dQuote{treatment} can only be \dQuote{control} and 
\dQuote{enriched}}

\item{circ_sg}{\code{SGSeq} object supplying feature info}

\item{test}{either \dQuote{DEX} for DEXSeq based feature selection or \dQuote{comparison} simple average
comaparison}
}
\value{
vector of featureID
}
\description{
CircRNA feature selection
}
\details{
This function works in 2 ways: direct comparison of average quantities or as a wrapper of DEXSeq.
In case of dataset with replicates, the suggested approach is the use of DEXSeq statistical test.
}
\author{
Stefan Stefanov
}
\keyword{depleted}
