% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{get.seqs}
\alias{get.seqs}
\title{Extract sequence per exon bin}
\usage{
get.seqs(full_sg, bs_genome = Dmelanogaster)
}
\arguments{
\item{full_sg}{a \code{SGSeq} object of exon bins}

\item{bs_genome}{\dQuote{BSgenome} name}
}
\value{
sequence list
}
\description{
A wrapper function for BSgenome subsequencing
}
\details{
Extracts sequence based on \code{SGSeq} object and \dQuote{BSgenome} name
}
\author{
Stefan Stefanov
}
\keyword{seqs}
