% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rides.R
\name{import_rides}
\alias{import_rides}
\title{Import multiple surveys}
\usage{
import_rides(
  eval_directory,
  filetype = "csv",
  file_pattern = "\\\\.csv",
  rename_cols = NULL,
  check_names = TRUE
)
}
\arguments{
\item{eval_directory}{the top level directory folder}

\item{filetype}{"csv". Currently only csv extensions are supported.}

\item{file_pattern}{a quoted regex expression to match file patterns in the directory. Extensions must be .csv.}

\item{rename_cols}{a vector of new names passed on to \code{data.table::fread()} to rename columns.}

\item{check_names}{TRUE a logical vector passed on to \code{data.table::fread()} to check raw column names.}
}
\value{
a dataframe of all appended survey files
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{import_rides} imports all rideshare files in a given directory and
outputs an appended dataframe
}
