% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stations.R
\name{import_stations}
\alias{import_stations}
\title{Import unique bikeshare stations}
\usage{
import_stations(
  eval_directory,
  filetype = "csv",
  file_pattern = "\\\\.csv",
  station_cols = tidyselect::contains("station"),
  gps_cols = tidyselect::contains(c("lat", "lng")),
  create_id = TRUE,
  remove_missings = TRUE,
  missing_chr = "",
  rename_expr = rlang::expr(c(station_name = tidyselect::ends_with(c("station",
    "name")), station_no = tidyselect::ends_with(c("number", "id")))),
  distinct_expr = rlang::expr(c(tidyselect::contains("station"))),
  convert_expr = rlang::expr(hablar::chr(station_no)),
  check_names = TRUE
)
}
\arguments{
\item{eval_directory}{the top level directory folder}

\item{filetype}{"csv". Currently only csv extensions are supported.}

\item{file_pattern}{a quoted regex expression to match file patterns in the directory. Extensions must be .csv.}

\item{station_cols}{a list of possible columns to check for station data. Defaults to tidyselect::contains("station")}

\item{gps_cols}{a tidyselect expression to determine the group of gps columns}

\item{create_id}{TRUE, create a project ID column after import?}

\item{remove_missings}{TRUE should missing character strings be removed?}

\item{missing_chr}{what character is equivalent to a missing value?}

\item{rename_expr}{an expression of \code{new_name} = \code{old_name} for columns to rename.}

\item{distinct_expr}{an expression to determine variables that define unique station name-number combinations.
Defaults to search across all variables resulting from the \code{rename_expr}.}

\item{convert_expr}{an expression to convert/ensure variable types. See ?hablar::convert()}

\item{check_names}{TRUE a logical vector passed on to \code{data.table::fread()} to check raw column names.}
}
\value{
a dataframe of all appended survey files
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{import_stations} imports station info from a directory of files and appends all unique
number combinations
}
