% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_year.R
\name{import_year}
\alias{import_year}
\title{Import data in year format}
\usage{
import_year(file, rename.list = NULL, na.strings = "", col.types = NULL)
}
\arguments{
\item{file}{The path to the file to import}

\item{rename.list}{A named vector of new names from old imported names}

\item{na.strings}{Passed on to \code{fread()}; how to interpret missing cells}

\item{col.types}{Passed on to colClasses of \code{fread()}; named or unnamed vector to specify col types.}
}
\value{
An imported file as a tibble
}
\description{
Import data in year format
}
