% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{make.BSJ.gr}
\alias{make.BSJ.gr}
\title{Convert BSJ string to GRanges obejct}
\usage{
make.BSJ.gr(BSJ_set)
}
\arguments{
\item{BSJ_set}{a list of BSJ ID records procudes by \code{process.BSJs} or \code{combine.two.BSJ.tables}}
}
\value{
\code{GRanges} object indicating BSJ loci
}
\description{
Convert BSJ string to GRanges obejct
}
\details{
Convert BSJ string to \code{GRanges} obejct
}
\author{
Stefan Stefanov
}
\keyword{BSJ}
\keyword{GRanges}
