% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{make_BSJ_gr}
\alias{make_BSJ_gr}
\title{Convert BSJ string to GRanges obejct}
\usage{
make_BSJ_gr(BSJ_set)
}
\arguments{
\item{BSJ_set}{a list of BSJ ID records procudes by \code{process_BSJs} or \code{combine_two_BSJ_tables}}
}
\value{
\code{GRanges} object indicating BSJ loci
}
\description{
Convert BSJ string to GRanges obejct
}
\details{
Convert BSJ string to \code{GRanges} obejct
}
\author{
Stefan Stefanov
}
\keyword{BSJ}
\keyword{GRanges}
