% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{make_BSJ_sg}
\alias{make_BSJ_sg}
\title{Preparation of the BSJ-specific splice graphs}
\usage{
make_BSJ_sg(circ_sg, BSJ_gr)
}
\arguments{
\item{circ_sg}{\code{SGSeq} prediction object}

\item{BSJ_gr}{a GRange of BSJ coordinates}
}
\value{
\code{SGSeq} containing exons belonging to BSJ loci
}
\description{
Selection of the exons based on BSJ set
}
\details{
Selection of the exons based on BSJ set
}
\author{
Stefan Stefanov
}
\keyword{GRanges}
\keyword{overlap}
