% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{merge_fasta}
\alias{merge_fasta}
\title{Merging 2 \code{DNAStringSets}}
\usage{
merge_fasta(qics_fa, known_fa)
}
\arguments{
\item{qics_fa}{\code{DNAStringSet} 1}

\item{known_fa}{\code{DNAStringSet} 2}
}
\value{
merged \code{DNAStringSet} file to be used for quantification
}
\description{
Merging 2 DNAStringSets
Merging 2 DNAStringSets
}
\author{
Stefan Stefanov
}
\keyword{FASTA}
\keyword{merge}
