% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{overlap.SG.BSJ}
\alias{overlap.SG.BSJ}
\title{Overlap of BSJ and a splice graph}
\usage{
overlap.SG.BSJ(sgfc_pred, BSJ_gr)
}
\arguments{
\item{sgfc_pred}{\code{SGSeq} prediction object}

\item{BSJ_gr}{a GRange of BSJ cooredinates}
}
\value{
\code{SGSeq} with disjoint exon bins
}
\description{
Creates a disjointed set of exons based on a \code{SGSeq} obejct and a BSJ \code{GRanges} object
}
\details{
Creates a disjointed set of exons based on a \code{SGSeq} obejct and a BSJ \code{GRanges} object. 
 The function keeps the \code{SGSeq} metadata
}
\author{
Stefan Stefanov
}
\keyword{GRanges}
\keyword{overlap}
