% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{overlap_SG_BSJ}
\alias{overlap_SG_BSJ}
\title{Overlap of BSJ and a splice graph}
\usage{
overlap_SG_BSJ(sgfc_pred, BSJ_gr, sg_annot)
}
\arguments{
\item{sgfc_pred}{\code{SGSeq} prediction object}

\item{BSJ_gr}{a GRange of BSJ cooredinates}

\item{sg_annot}{\code{SGSeq} prediction object with annotation info}
}
\value{
\code{SGSeq} with disjoint exon bins
}
\description{
Creates a disjointed set of exons based on a \code{SGSeq} obejct and a BSJ \code{GRanges} object
}
\details{
Creates a disjointed set of exons based on a \code{SGSeq} obejct and a BSJ \code{GRanges} object. 
 The function keeps the \code{SGSeq} metadata
}
\author{
Stefan Stefanov
}
\keyword{GRanges}
\keyword{overlap}
