% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{parse.files}
\alias{parse.files}
\title{Parse BSJ input}
\usage{
parse.files(file_list, file_path, input_type)
}
\arguments{
\item{file_list}{list with file names}

\item{file_path}{string object with file path, clould be an empty string}

\item{input_type}{\code{CIRI} for CIRI2 input, \code{CE} for CIRCexplorer2 input and \code{tsv} 
for TSV formatted input}
}
\value{
\code{Tibble} object with combined BSJ coordinate and number of junction spanning 
 reads across sample
}
\description{
Parse BSJ files from CIRI, CIRCexplorer2 or a TSV file
}
\details{
This processes BSJ prediction files and prepares them for the next step of the pipeline.
\code{input_type} is essential for the correct parsing of the files.
}
\author{
Stefan Stefanov
}
\keyword{parse}
