% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{plotRanges2}
\alias{plotRanges2}
\title{Plot ranges}
\usage{
plotRanges2(...)
}
\value{
ggplot of multiple GRanges objects
}
\description{
Plots GRanges objects
}
\details{
ggplot of multiple GRanges object. Every object is auto assigned a colour from colorblind friendly scheme
}
\author{
Stefan Stefanov
}
\keyword{GRanges}
\keyword{plot}
