% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{prep.output.gtf}
\alias{prep.output.gtf}
\title{Creation of GTF based on CYCLeR results}
\usage{
prep.output.gtf(qics_out, circ_exons, annot_list = NULL)
}
\arguments{
\item{circ_exons}{\code{SGSeq} object supplying feature info}

\item{annot_list}{ORG package; soon to be expanded}

\item{qics}{CYCLeR table  of intermediate results}
}
\value{
GTF-like table
}
\description{
Creation of GTF based on CYCLeR results
}
\details{
This function takes the
}
\author{
Stefan Stefanov
}
\keyword{GTF}
