% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{prep_circular_sg}
\alias{prep_circular_sg}
\title{Preparing circular splice graph features}
\usage{
prep_circular_sg(full_sg, full_fc, sgfc_pred, bs_genome, BSJ_gr, th = 15)
}
\arguments{
\item{full_sg}{\code{SGranges} object supplying feature info}

\item{full_fc}{count matrix corresponding to the features}

\item{sgfc_pred}{\code{SGFC} from the SGSeq feature detection}

\item{bs_genome}{\code{BSGenome} object used for extracting the sequences}

\item{BSJ_gr}{\code{GRange} object of BSJ coordinates}

\item{th}{minimum number of reads limiting exon selection}
}
\value{
\code{SummarizedExperiment} object with info on the circular splice graph features
}
\description{
Preparing circular splice graph features
Preparing circular splice graph features
}
\author{
Stefan Stefanov
}
\keyword{graph}
\keyword{preparation}
\keyword{splice}
