% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{prep_output_gtf}
\alias{prep_output_gtf}
\title{Creation of GTF based on CYCLeR results}
\usage{
prep_output_gtf(qics_out, sgfc, annot_list = NULL)
}
\arguments{
\item{sgfc}{\code{SummarizedExperiment} object with info on the circular splice graph features}

\item{annot_list}{ORG package; soon to be expanded}

\item{qics}{CYCLeR \code{data.frame}  of intermediate results}
}
\value{
GTF-like table
}
\description{
Creation of GTF based on CYCLeR results
}
\details{
This function takes the
}
\author{
Stefan Stefanov
}
\keyword{GTF}
