% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{process_BSJs}
\alias{process_BSJs}
\title{Process BSJs}
\usage{
process_BSJs(cdf, sample_table)
}
\arguments{
\item{cdf}{tibble produced by \code{parse_files}}

\item{sample_table}{sample table formatted according to the manual,
Must contain \dQuote{sample_name} \dQuote{treatment} \dQuote{file_bam} \dQuote{lib_size} 
\dQuote{read_len}; NB the values in column \dQuote{treatment} can only be \dQuote{control} and 
\dQuote{enriched}}

\item{file_path}{string object with file path, clould be an empty string}
}
\value{
\code{Tibble} object with combined filtered BSJ coordinate and number of junction spanning 
 reads across sample.
}
\description{
process the BSJ table and select high confidence BSJs
}
\details{
Filters BSJ based on comparison of the average CPM values of BSJs
}
\author{
Stefan Stefanov
}
\keyword{BSJ}
\keyword{filter}
