% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{recount.features}
\alias{recount.features}
\title{Re-count of the reads per exon bin}
\usage{
recount.features(full_sg, sample_table)
}
\arguments{
\item{full_sg}{a \code{SGSeq} obeject of exon bins}

\item{sample_table}{sample table formatted according to the manual,
Must contain \dQuote{sample_name} \dQuote{treatment} \dQuote{file_bam} \dQuote{lib_size} 
\dQuote{read_len}; NB the values in column \dQuote{treatment} can only be \dQuote{control} and 
\dQuote{enriched}}
}
\value{
\code{BAMFileList} object with info on the trimmed files
}
\description{
A wrapper function for Rsubread
}
\details{
This function performs requantification of the exon bins with specifically selected parameters
}
\author{
Stefan Stefanov
}
\keyword{Bam}
\keyword{Filter}
