% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{RPKM.calc}
\alias{RPKM.calc}
\title{RPKM calculation for the genomic features}
\usage{
RPKM.calc(
  count_matrix,
  sg,
  bsj_granges,
  bs_genome,
  sample_table,
  feature_type,
  fsj_overhang = 3,
  bsj_overhang = 15,
  eff_length_correction = T,
  gc_correction = F
)
}
\arguments{
\item{count_matrix}{count matrix corresponding to the features}

\item{sg}{\code{SGSeq} object supplying feature info}

\item{bsj_granges}{GRange of BSJ cooredinates}

\item{bs_genome}{a \code{BSGenome} object used for extracting the sequences}

\item{sample_table}{sample table formatted according to the manual,
Must contain \dQuote{sample_name} \dQuote{treatment} \dQuote{file_bam} \dQuote{lib_size} 
\dQuote{read_len}; NB the values in column \dQuote{treatment} can only be \dQuote{control} and 
\dQuote{enriched}}

\item{feature_type}{either \dQuote{e} for exons ot \dQuote{j} for junctions}

\item{fsj_overhang}{the FJS overhand used in the mapping a.k.a. anchor}

\item{bsj_overhang}{the BSJ overhand used in the chimeric detection}

\item{eff_length_correction}{whether or not to apply effective length correction}

\item{gc_correction}{whether or not to apply GC-content correction; requires further testing}
}
\value{
\code{BAMFileList} object with info on the trimmed files
}
\description{
RPKM calculation for the genomic features
}
\details{
This function performs RPKM calculations for the exonic features. The RPKM calculation is
 performed based on the exact sequences for the exons. For junctions, the sequences are selected based 
 on the exons, flanking the junction. The function takes into account the needed effective 
 length corrections.
}
\author{
Stefan Stefanov
}
\keyword{RPKM}
