% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_time.R
\name{segment_time}
\alias{segment_time}
\title{Segmentation of the cycling route according to activity}
\usage{
segment_time(gps)
}
\arguments{
\item{gps}{the GPS data of the cycling route (longitude, latitude, time, elevation, ...)}
}
\value{
A list containing:
 \itemize{
   \item moving_ind indexes of GPS data points with velocity greater than 1 m/s,
   \item resting_ind indexes of GPS data points with velocity below 1 m/s.
   \item time_div data frame
     \itemize{
       \item total_time total duration of the cycling route,
       \item moving_time duration of the cyclist traveling faster than 1 m/s,
       \item t_mov_rel moving time relative to total time,
       \item resting_time duration of the cyclist traveling slower than 1 m/s,
       \item t_rest_rel resting time relative to total time,
       \item time_asc total time spent ascending,
       \item t_asc_rel time spent ascending relative to total time,
       \item time_desc total time spent descending,
       \item t_desc_rel time spent descending relative to total time,
       \item time_flat total time spent on flat,
       \item t_flat_rel time spent on flat relative to total time,
       \item time_NA total time with undefined (NA) grade because of smoothing,
       \item t_NA_rel time with undefined (NA) grade because of smoothing relative to total time.
     }
 }
}
\description{
This function determines the total time, moving time and resting time on a cycling route described with GPS data.
Additionally, the function calculates the time spent ascending, descending and on the flat.
}
\examples{
load(system.file("extdata", "gps.Rda", package="cycleR"))
segments <- segment_time(gps)
}
\keyword{GPS}
\keyword{analysis,}
\keyword{cycling,}
\keyword{data}
\keyword{moving}
\keyword{time}

