% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_stations.R
\name{split_stations}
\alias{split_stations}
\title{Removes station information and tables it for later.}
\usage{
split_stations(
  df,
  start_name = "start",
  start_no = "startno",
  end_name = "end",
  end_no = "endno",
  return_list = FALSE,
  preserve_numbers = TRUE
)
}
\arguments{
\item{df}{the dataframe that contains ride-level data.}

\item{start_name}{a character vector of names of columns that may contain start name data}

\item{start_no}{a character vector of names of columns that may contain start number data}

\item{end_name}{a character vector of names of columns that may contain end name data}

\item{end_no}{a character vector of names of columns that may contain end number data}

\item{return_list}{FALSE, should the return object include the removed station info?}

\item{preserve_numbers}{TRUE, should the columns containing the start and end station numbers be maintained?}
}
\value{
a list object of two dataframes: the ride-level data with station info removed and the
unique station name-number combinations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{split_stations} removes station variables from a ride-level data-frame and returns a
list object containing the resulting trimmed dataframe and a second dataframe that contains
the unique station info that was extracted.
}
