% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{transcripts.per.sample}
\alias{transcripts.per.sample}
\title{Transcript assembly}
\usage{
transcripts.per.sample(i)
}
\arguments{
\item{i}{name of the sample}
}
\value{
\code{data.frame} of transcript information in flat format
}
\description{
Transcript assembly per sample 
Transcript assembly per sample based on sample name in the \dQuote{sample_table}
}
\author{
Stefan Stefanov
}
\keyword{assembly}
