% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycler_fuctions.R
\name{transcripts_per_sample}
\alias{transcripts_per_sample}
\title{Transcript assembly}
\usage{
transcripts_per_sample(sgfc, BSJ_gr, sample_name)
}
\arguments{
\item{sgfc}{\code{SummarizedExperiment} object with info on the circular splice graph features}

\item{BSJ_gr}{\code{GRange} object of BSJ coordinates}

\item{sample_name}{name of the sample as indicated in the sample table}
}
\value{
\code{data.frame} of transcript information in flat format
}
\description{
Transcript assembly per sample 
Transcript assembly per sample based on sample name in the \dQuote{sample_table}
}
\author{
Stefan Stefanov
}
\keyword{assembly}
