% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_od.R
\name{transform_od}
\alias{transform_od}
\title{Transform ride-level data into OD data}
\usage{
transform_od(
  df,
  origin_datetime,
  destination_datetime,
  timezone = "US/Eastern",
  weekstart = 7,
  member_var,
  station_vars = c("startno", "endno"),
  round_dig = 3,
  granularity = NULL
)
}
\arguments{
\item{df}{the dataframe that contains ride-level data.}

\item{origin_datetime}{the column containing origin/departure date-time information.}

\item{destination_datetime}{the column containing destination/arrival date-time information.}

\item{timezone}{"US/Eastern", the timezone passed to \code{lubridate}. See \code{base::OlsonNames()} for valid tz names.}

\item{weekstart}{7, an integer passed to \code{base::getOption('lubridate.week.start')}. Determines the start day of the week. 7 = Sunday.}

\item{member_var}{member, an unquoted name of the variable containing the string member data.}

\item{station_vars}{a character vector of variable names that define the unique stations.}

\item{round_dig}{3, the number of digits when rounding.}

\item{granularity}{NULL one of "hour", "day", or "month". The precision to which the data will be collapsed.}
}
\value{
a dataframe of summarized origin-destination information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{transform_od} takes ride-level data in a data.frame and transforms the data into
origin-destination data by summarizing at a user-defined time interval
}
