% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_colnames.R
\name{view_colnames}
\alias{view_colnames}
\title{Get an overview of all csv column names in a directory}
\usage{
view_colnames(
  eval_directory,
  filetype = "csv",
  file_pattern = "\\\\.csv",
  n_rows = 10,
  check_names = FALSE
)
}
\arguments{
\item{eval_directory}{the top level directory folder}

\item{filetype}{"csv". Currently only csv extensions are supported.}

\item{file_pattern}{a quoted regex expression to match file patterns in the directory. Extensions must be .csv.}

\item{n_rows}{10, the number of rows to import for each csv file.}

\item{check_names}{FALSE a logical vector passed on to \code{data.table::fread()} to check raw column names.}
}
\value{
a dataframe of all appended survey files
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{view_colnames} imports the first n rows of each csv file in a given directory and reports column names
}
