% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYCLeR_functions.R
\name{merge_qics}
\alias{merge_qics}
\title{Merging 2 assemblies}
\usage{
merge_qics(qics1, qics2)
}
\arguments{
\item{qics1}{assembly 1}

\item{qics2}{assembly 2}
}
\value{
\code{data.frame} of transcript information in flat format
}
\description{
Pair-wise merging 2 assemblies
Pair-wise merging 2 assemblies
}
\author{
Stefan Stefanov
}
\keyword{assembly}
